+++ src/bitcoin.rs
@@ -1,0 +1,308 @@
+use std::str::FromStr;
+
+pub use ::bitcoin::util::address::Address;
+use anyhow::Result;
+pub use bdk::{wallet::AddressIndex, FeeRate, LocalUtxo, TransactionDetails};
+pub use bitcoin_hashes::{sha256, Hash};
+use serde_encrypt::{
+    serialize::impls::BincodeSerializer, shared_key::SharedKey, traits::SerdeEncryptSharedKey,
+    AsSharedKey, EncryptedMessage,
+};
+use tokio::try_join;
+
+mod assets;
+mod keys;
+mod payment;
+mod psbt;
+mod wallet;
+
+pub use crate::{
+    bitcoin::{
+        assets::dust_tx,
+        keys::{new_mnemonic, save_mnemonic},
+        payment::{create_payjoin, create_transaction},
+        psbt::sign_psbt,
+        wallet::{get_blockchain, get_wallet, synchronize_wallet},
+    },
+    debug, info,
+    structs::{
+        EncryptedWalletData, FundVaultDetails, MnemonicSeedData, SatsInvoice, WalletData,
+        WalletTransaction,
+    },
+    trace,
+};
+
+impl SerdeEncryptSharedKey for EncryptedWalletData {
+    type S = BincodeSerializer<Self>; // you can specify serializer implementation (or implement it by yourself).
+}
+
+/// Bitcoin Wallet Operations
+
+pub fn get_encrypted_wallet(
+    password: &str,
+    encrypted_descriptors: &str,
+) -> Result<EncryptedWalletData> {
+    // read hash digest and consume hasher
+    let hash = sha256::Hash::hash(password.as_bytes());
+    let shared_key: [u8; 32] = hash.into_inner();
+    let encrypted_descriptors: Vec<u8> = hex::decode(encrypted_descriptors)?;
+    let encrypted_message = EncryptedMessage::deserialize(encrypted_descriptors)?;
+    Ok(EncryptedWalletData::decrypt_owned(
+        &encrypted_message,
+        &SharedKey::from_array(shared_key),
+    )?)
+}
+
+pub async fn new_mnemonic_seed(
+    encryption_password: &str,
+    seed_password: &str,
+) -> Result<MnemonicSeedData> {
+    let hash = sha256::Hash::hash(encryption_password.as_bytes());
+    let shared_key: [u8; 32] = hash.into_inner();
+
+    let encrypted_wallet_data = new_mnemonic(seed_password).await?;
+    let encrypted_message = encrypted_wallet_data.encrypt(&SharedKey::from_array(shared_key))?;
+    let serialized_encrypted_message = hex::encode(encrypted_message.serialize());
+    let mnemonic_seed_data = MnemonicSeedData {
+        mnemonic: encrypted_wallet_data.private.mnemonic,
+        serialized_encrypted_message,
+    };
+
+    Ok(mnemonic_seed_data)
+}
+
+pub async fn save_mnemonic_seed(
+    mnemonic_phrase: &str,
+    encryption_password: &str,
+    seed_password: &str,
+) -> Result<MnemonicSeedData> {
+    let hash = sha256::Hash::hash(encryption_password.as_bytes());
+    let shared_key: [u8; 32] = hash.into_inner();
+
+    let vault_data = save_mnemonic(mnemonic_phrase, seed_password).await?;
+    let encrypted_message = vault_data.encrypt(&SharedKey::from_array(shared_key))?;
+    let serialized_encrypted_message = hex::encode(encrypted_message.serialize());
+    let mnemonic_seed_data = MnemonicSeedData {
+        mnemonic: vault_data.private.mnemonic,
+        serialized_encrypted_message,
+    };
+
+    Ok(mnemonic_seed_data)
+}
+
+pub async fn get_wallet_data(
+    descriptor: &str,
+    change_descriptor: Option<String>,
+) -> Result<WalletData> {
+    info!("get_wallet_data");
+    info!(format!("descriptor: {descriptor}"));
+    info!(format!("change_descriptor {change_descriptor:?}"));
+
+    let wallet = get_wallet(descriptor, change_descriptor).await?;
+    synchronize_wallet(&wallet).await?;
+    let address = wallet.get_address(AddressIndex::LastUnused)?.to_string();
+    info!(format!("address: {address}"));
+    let balance = wallet.get_balance()?;
+    info!(format!("balance: {balance:?}"));
+    let utxos = wallet.list_unspent().unwrap_or_default();
+    let utxos: Vec<String> = utxos.into_iter().map(|x| x.outpoint.to_string()).collect();
+    trace!(format!("unspent: {utxos:#?}"));
+
+    let mut transactions = wallet.list_transactions(false).unwrap_or_default();
+    trace!(format!("transactions: {transactions:#?}"));
+
+    transactions.sort_by(|a, b| {
+        b.confirmation_time
+            .as_ref()
+            .map(|t| t.height)
+            .cmp(&a.confirmation_time.as_ref().map(|t| t.height))
+    });
+
+    let transactions: Vec<WalletTransaction> = transactions
+        .into_iter()
+        .map(|tx| WalletTransaction {
+            txid: tx.txid,
+            received: tx.received,
+            sent: tx.sent,
+            fee: tx.fee,
+            confirmed: tx.confirmation_time.is_some(),
+            confirmation_time: tx.confirmation_time,
+        })
+        .collect();
+
+    Ok(WalletData {
+        address,
+        balance,
+        transactions,
+        utxos,
+    })
+}
+
+pub async fn get_new_address(
+    descriptor: &str,
+    change_descriptor: Option<String>,
+) -> Result<String> {
+    info!("get_new_address");
+    info!(format!("descriptor: {descriptor}"));
+    info!(format!("change_descriptor: {change_descriptor:?}"));
+
+    let wallet = get_wallet(descriptor, change_descriptor).await?;
+    synchronize_wallet(&wallet).await?;
+    let address = wallet.get_address(AddressIndex::New)?.to_string();
+    info!(format!("address: {address}"));
+    Ok(address)
+}
+
+pub async fn send_sats(
+    descriptor: &str,
+    change_descriptor: &str,
+    destination: &str, // bip21 uri or address
+    amount: u64,
+    fee_rate: Option<f32>,
+) -> Result<TransactionDetails> {
+    use payjoin::UriExt;
+
+    let wallet = get_wallet(descriptor, Some(change_descriptor.to_owned())).await?;
+    synchronize_wallet(&wallet).await?;
+
+    let fee_rate = fee_rate.map(FeeRate::from_sat_per_vb);
+
+    let transaction = match payjoin::Uri::try_from(destination) {
+        Ok(uri) => {
+            let address = uri.address.clone();
+            if let Ok(pj_uri) = uri.check_pj_supported() {
+                create_payjoin(
+                    vec![SatsInvoice { address, amount }],
+                    &wallet,
+                    fee_rate,
+                    pj_uri,
+                )
+                .await?
+            } else {
+                create_transaction(vec![SatsInvoice { address, amount }], &wallet, fee_rate).await?
+            }
+        }
+        _ => {
+            let address = Address::from_str(destination)?;
+            create_transaction(vec![SatsInvoice { address, amount }], &wallet, fee_rate).await?
+        }
+    };
+
+    Ok(transaction)
+}
+
+pub async fn fund_vault(
+    btc_descriptor_xprv: &str,
+    btc_change_descriptor_xprv: &str,
+    assets_address: &str,
+    uda_address: &str,
+    asset_amount: u64,
+    uda_amount: u64,
+    fee_rate: Option<f32>,
+) -> Result<FundVaultDetails> {
+    let assets_address = Address::from_str(assets_address)?;
+    let uda_address = Address::from_str(uda_address)?;
+
+    let wallet = get_wallet(
+        btc_descriptor_xprv,
+        Some(btc_change_descriptor_xprv.to_owned()),
+    )
+    .await?;
+    synchronize_wallet(&wallet).await?;
+
+    let asset_invoice = SatsInvoice {
+        address: assets_address,
+        amount: asset_amount,
+    };
+    let uda_invoice = SatsInvoice {
+        address: uda_address,
+        amount: uda_amount,
+    };
+
+    let fee_rate = fee_rate.map(FeeRate::from_sat_per_vb);
+
+    let asset_tx_details = create_transaction(
+        vec![asset_invoice.clone(), asset_invoice],
+        &wallet,
+        fee_rate,
+    )
+    .await?;
+
+    let uda_tx_details =
+        create_transaction(vec![uda_invoice.clone(), uda_invoice], &wallet, fee_rate).await?;
+
+    let asset_txid = asset_tx_details.txid;
+    let asset_outputs: Vec<String> = asset_tx_details
+        .transaction
+        .expect("asset tx exists")
+        .output
+        .iter()
+        .enumerate()
+        .map(|(i, _)| format!("{asset_txid}:{i}"))
+        .collect();
+
+    let uda_txid = uda_tx_details.txid;
+    let uda_outputs: Vec<String> = uda_tx_details
+        .transaction
+        .expect("uda tx exists")
+        .output
+        .iter()
+        .enumerate()
+        .map(|(i, _)| format!("{uda_txid}:{i}"))
+        .collect();
+
+    Ok(FundVaultDetails {
+        assets_output: Some(asset_outputs[0].to_owned()),
+        assets_change_output: Some(asset_outputs[1].to_owned()),
+        udas_output: Some(uda_outputs[0].to_owned()),
+        udas_change_output: Some(uda_outputs[1].to_owned()),
+        is_funded: true,
+    })
+}
+
+fn utxo_string(utxo: &LocalUtxo) -> String {
+    utxo.outpoint.to_string()
+}
+
+pub async fn get_assets_vault(
+    rgb_assets_descriptor_xpub: &str,
+    rgb_udas_descriptor_xpub: &str,
+) -> Result<FundVaultDetails> {
+    let assets_wallet = get_wallet(rgb_assets_descriptor_xpub, None).await?;
+    let udas_wallet = get_wallet(rgb_udas_descriptor_xpub, None).await?;
+
+    try_join!(
+        synchronize_wallet(&assets_wallet),
+        synchronize_wallet(&udas_wallet)
+    )?;
+
+    let assets_utxos = assets_wallet.list_unspent()?;
+    let uda_utxos = udas_wallet.list_unspent()?;
+
+    debug!(format!("Asset UTXOs: {assets_utxos:#?}"));
+    debug!(format!("UDA UTXOs: {uda_utxos:#?}"));
+
+    let mut assets_utxos: Vec<String> = assets_utxos.iter().map(utxo_string).collect();
+    assets_utxos.sort();
+
+    let mut uda_utxos: Vec<String> = uda_utxos.iter().map(utxo_string).collect();
+    uda_utxos.sort();
+
+    let assets_change_output = assets_utxos.pop();
+    let assets_output = assets_utxos.pop();
+    let udas_change_output = uda_utxos.pop();
+    let udas_output = uda_utxos.pop();
+
+    let is_funded = assets_change_output.is_some()
+        && assets_output.is_some()
+        && udas_change_output.is_some()
+        && udas_output.is_some();
+
+    Ok(FundVaultDetails {
+        assets_output,
+        assets_change_output,
+        udas_output,
+        udas_change_output,
+        is_funded,
+    })
+}
