+++ src/bitcoin/assets.rs
