+++ src/constants.rs
@@ -8,7 +8,7 @@ use tokio::sync::RwLock;
 use crate::error;
 
 fn dot_env(key: &str) -> String {
-    let env_file = include_str!("../../.env");
+    let env_file = include_str!("../.env");
     match env::var(key) {
         Ok(val) => val,
         Err(_) => {
