+++ src/data.rs
@@ -1,2 +1,1 @@
-pub mod constants;
 pub mod structs;
