--- src/operations.rs
@@ -1,4 +1,0 @@
-pub mod bitcoin;
-pub mod carbonado;
-pub mod lightning;
-pub mod rgb;
