--- src/operations/bitcoin.rs
@@ -1,11 +1,0 @@
-mod assets;
-mod balance;
-mod psbt;
-mod secret;
-mod send_sats;
-
-pub use self::psbt::sign_psbt;
-pub use assets::dust_tx;
-pub use balance::{get_wallet, synchronize_wallet};
-pub use secret::{new_mnemonic, save_mnemonic};
-pub use send_sats::{create_payjoin, create_transaction};
