--- src/operations/rgb.rs
@@ -1,8 +1,0 @@
-pub mod accept;
-pub mod constants;
-pub mod invoice;
-pub mod issue;
-pub mod psbt;
-pub mod resolvers;
-pub mod schemas;
-mod structs;
