+++ src/rgb/constants.rs
