+++ src/rgb/schemas.rs
