+++ src/rgb/structs.rs
